+++ common/goos/Reader.h
@@ -73,7 +73,7 @@ class Reader {
   Object read_from_string(const std::string& str, bool add_top_level = true);
   std::optional<Object> read_from_stdin(const std::string& prompt, ReplWrapper& repl);
   Object read_from_file(const std::vector<std::string>& file_path);
-
+  bool check_string_is_valid(const std::string& str) const;
   std::string get_source_dir();
 
   SymbolTable symbolTable;
@@ -97,9 +97,10 @@ class Reader {
   bool read_string(TextStream& stream, Object& obj);
   void add_reader_macro(const std::string& shortcut, std::string replacement);
 
-  char valid_symbols_chars[256];
+  bool m_valid_symbols_chars[256];
+  bool m_valid_source_text_chars[256];
 
-  std::unordered_map<std::string, std::string> reader_macros;
+  std::unordered_map<std::string, std::string> m_reader_macros;
 };
 
 std::string get_readable_string(const char* in);