+++ decompiler/IR2/AtomicOp.cpp
@@ -7,6 +7,7 @@
 #include "decompiler/ObjectFile/LinkedObjectFile.h"
 #include "AtomicOp.h"
 #include "OpenGoalMapping.h"
+#include "Form.h"
 
 namespace decompiler {
 /////////////////////////////
@@ -291,6 +292,14 @@ std::string get_simple_expression_op_name(SimpleExpression::Kind kind) {
       return "max.ui";
     case SimpleExpression::Kind::PCPYLD:
       return "pcypld";
+    case SimpleExpression::Kind::VECTOR_PLUS:
+      return "vector+!2";
+    case SimpleExpression::Kind::VECTOR_MINUS:
+      return "vector-!2";
+    case SimpleExpression::Kind::VECTOR_FLOAT_PRODUCT:
+      return "vector-float*!2";
+    case SimpleExpression::Kind::SUBU_L32_S7:
+      return "subu-s7";
     default:
       assert(false);
       return {};
@@ -342,6 +351,12 @@ int get_simple_expression_arg_count(SimpleExpression::Kind kind) {
     case SimpleExpression::Kind::MAX_UNSIGNED:
     case SimpleExpression::Kind::PCPYLD:
       return 2;
+    case SimpleExpression::Kind::VECTOR_PLUS:
+    case SimpleExpression::Kind::VECTOR_MINUS:
+    case SimpleExpression::Kind::VECTOR_FLOAT_PRODUCT:
+      return 3;
+    case SimpleExpression::Kind::SUBU_L32_S7:
+      return 1;
     default:
       assert(false);
       return -1;
@@ -360,6 +375,18 @@ SimpleExpression::SimpleExpression(Kind kind, const SimpleAtom& arg0, const Simp
   m_args[1] = arg1;
   m_kind = kind;
   assert(get_simple_expression_arg_count(kind) == 2);
+}
+
+SimpleExpression::SimpleExpression(Kind kind,
+                                   const SimpleAtom& arg0,
+                                   const SimpleAtom& arg1,
+                                   const SimpleAtom& arg2)
+    : n_args(3) {
+  m_args[0] = arg0;
+  m_args[1] = arg1;
+  m_args[2] = arg2;
+  m_kind = kind;
+  assert(get_simple_expression_arg_count(kind) == 3);
 }
 
 goos::Object SimpleExpression::to_form(const std::vector<DecompilerLabel>& labels,