+++ decompiler/IR2/FormExpressionAnalysis.cpp
@@ -129,6 +129,39 @@ bool is_power_of_two(int in, int* out) {
 }
 
 /*!
+ * Imagine:
+ *   x = foo
+ *   { // some macro/inlined thing
+ *     read from x
+ *     return x;
+ *   }
+ *
+ * and you want to transform it to
+ * x = some_macro(foo, blah, ...)
+ *
+ * this will get you foo (and pop it from the stack), assuming the stack is sitting right after the
+ * point where the inline thing evaluated foo.
+ *
+ * For later book-keeping of reg use, if it gets you something new, it will set found_orig_out,
+ * and also give you the regaccess for the x of the x = foo.
+ *
+ * If you use this, you are responsible for adding code that sets x again.
+ */
+Form* repop_passthrough_arg(Form* in,
+                            FormStack& stack,
+                            const Env& env,
+                            RegisterAccess* orig_out,
+                            bool* found_orig_out) {
+  *found_orig_out = false;
+
+  auto as_atom = form_as_atom(in);
+  if (as_atom && as_atom->is_var()) {
+    return stack.pop_reg(as_atom->var().reg(), {}, env, true, -1, orig_out, found_orig_out);
+  }
+  return in;
+}
+
+/*!
  * Create a form which represents a variable.
  */
 Form* var_to_form(const RegisterAccess& var, FormPool& pool) {
@@ -181,15 +214,14 @@ void pop_helper(const std::vector<RegisterAccess>& vars,
           submit_reg_to_var.push_back(var_idx);
           submit_regs.push_back(var.reg());
         } else {
-          /*
-          auto var_id = env.get_program_var_id(var);
-          fmt::print(
-              "Unsafe to pop {}: used {} times, def {} times, expected use {} ({} {} rd: {}) ({} "
-              "{})\n",
-              var.to_string(env), use_def.use_count(), use_def.def_count(), times,
-              var.reg().to_string(), var.idx(), var.mode() == AccessMode::READ,
-              var_id.reg.to_string(), var_id.id);
-              */
+          // auto var_id = env.get_program_var_id(var);
+          //          fmt::print(
+          //              "Unsafe to pop {}: used {} times, def {} times, expected use {} ({} {} rd:
+          //              {}) ({} "
+          //              "{})\n",
+          //              var.to_string(env), use_def.use_count(), use_def.def_count(), times,
+          //              var.reg().to_string(), var.idx(), var.mode() == AccessMode::READ,
+          //              var_id.reg.to_string(), var_id.id);
 
           //          if (var.to_string(env) == "a3-0") {
           //            for (auto& use : use_def.uses) {
@@ -455,16 +487,13 @@ void SimpleExpressionElement::update_from_stack_identity(const Env& env,
       auto kv = env.label_types().find(lab.name);
       if (kv != env.label_types().end()) {
         auto type_name = kv->second.type_name;
+        // the actual decompilation is deferred until later, once static lambdas are done.
         if (type_name == "_auto_") {
-          auto decompiled_data = decompile_at_label_guess_type(lab, env.file->labels,
-                                                               env.file->words_by_seg, env.dts->ts);
-          result->push_back(pool.alloc_element<DecompiledDataElement>(decompiled_data));
+          result->push_back(pool.alloc_element<DecompiledDataElement>(lab));
         } else if (type_name == "_lambda_") {
           result->push_back(this);
         } else {
-          auto decompiled_data = decompile_at_label_with_hint(kv->second, lab, env.file->labels,
-                                                              env.file->words_by_seg, *env.dts);
-          result->push_back(pool.alloc_element<DecompiledDataElement>(decompiled_data));
+          result->push_back(pool.alloc_element<DecompiledDataElement>(lab, kv->second));
         }
       } else {
         result->push_back(this);
@@ -658,7 +687,7 @@ void SimpleExpressionElement::update_from_stack_add_i(const Env& env,
       input.stride = 1;
       input.base_type = arg1_type.typespec();
       auto out = env.dts->ts.reverse_field_lookup(input);
-      if (out.success) {
+      if (out.success && out.has_variable_token()) {
         // it is. now we have to modify things
         // first, look for the index
 
@@ -677,6 +706,7 @@ void SimpleExpressionElement::update_from_stack_add_i(const Env& env,
               tokens.push_back(to_token(tok));
             }
           }
+          assert(used_index);
           result->push_back(pool.alloc_element<DerefElement>(args.at(1), out.addr_of, tokens));
           return;
         } else {
@@ -690,7 +720,7 @@ void SimpleExpressionElement::update_from_stack_add_i(const Env& env,
       input.stride = arg0_type.get_mult_int_constant();
       input.base_type = arg1_type.typespec();
       auto out = env.dts->ts.reverse_field_lookup(input);
-      if (out.success) {
+      if (out.success && out.has_variable_token()) {
         // it is. now we have to modify things
         // first, look for the index
         int p2;
@@ -714,6 +744,7 @@ void SimpleExpressionElement::update_from_stack_add_i(const Env& env,
                 tokens.push_back(to_token(tok));
               }
             }
+            assert(used_index);
             result->push_back(pool.alloc_element<DerefElement>(args.at(1), out.addr_of, tokens));
             return;
           } else {
@@ -740,6 +771,7 @@ void SimpleExpressionElement::update_from_stack_add_i(const Env& env,
                 tokens.push_back(to_token(tok));
               }
             }
+            assert(used_index);
             result->push_back(pool.alloc_element<DerefElement>(args.at(1), out.addr_of, tokens));
             return;
           } else {
@@ -759,7 +791,7 @@ void SimpleExpressionElement::update_from_stack_add_i(const Env& env,
       rd_in.base_type = arg0_type.typespec();
       auto rd = env.dts->ts.reverse_field_lookup(rd_in);
 
-      if (rd.success) {
+      if (rd.success && rd.has_variable_token()) {
         auto arg1_matcher = Matcher::match_or(
             {Matcher::op(GenericOpMatcher::fixed(FixedOperatorKind::MULTIPLICATION),
                          {Matcher::any(0), Matcher::integer(rd_in.stride)}),
@@ -778,6 +810,7 @@ void SimpleExpressionElement::update_from_stack_add_i(const Env& env,
               tokens.push_back(to_token(tok));
             }
           }
+          assert(used_index);
           result->push_back(pool.alloc_element<DerefElement>(args.at(0), rd.addr_of, tokens));
           return;
         } else {
@@ -948,6 +981,54 @@ void SimpleExpressionElement::update_from_stack_pcypld(const Env& env,
 
   auto new_form = pool.alloc_element<GenericElement>(
       GenericOperator::make_fixed(FixedOperatorKind::PCPYLD), args.at(0), args.at(1));
+  result->push_back(new_form);
+}
+
+void SimpleExpressionElement::update_from_stack_vector_plus_minus(bool is_add,
+                                                                  const Env& env,
+                                                                  FormPool& pool,
+                                                                  FormStack& stack,
+                                                                  std::vector<FormElement*>* result,
+                                                                  bool allow_side_effects) {
+  std::vector<Form*> popped_args =
+      pop_to_forms({m_expr.get_arg(0).var(), m_expr.get_arg(1).var(), m_expr.get_arg(2).var()}, env,
+                   pool, stack, allow_side_effects);
+
+  for (int i = 0; i < 3; i++) {
+    auto arg_type = env.get_types_before_op(m_my_idx).get(m_expr.get_arg(i).var().reg());
+    if (arg_type.typespec() != TypeSpec("vector")) {
+      popped_args.at(i) = cast_form(popped_args.at(i), TypeSpec("vector"), pool, env);
+    }
+  }
+
+  auto new_form = pool.alloc_element<GenericElement>(
+      GenericOperator::make_fixed(is_add ? FixedOperatorKind::VECTOR_PLUS
+                                         : FixedOperatorKind::VECTOR_MINUS),
+      std::vector<Form*>{popped_args.at(0), popped_args.at(1), popped_args.at(2)});
+  result->push_back(new_form);
+}
+
+void SimpleExpressionElement::update_from_stack_vector_float_product(
+    const Env& env,
+    FormPool& pool,
+    FormStack& stack,
+    std::vector<FormElement*>* result,
+    bool allow_side_effects) {
+  std::vector<Form*> popped_args =
+      pop_to_forms({m_expr.get_arg(0).var(), m_expr.get_arg(1).var(), m_expr.get_arg(2).var()}, env,
+                   pool, stack, allow_side_effects);
+
+  for (int i = 0; i < 3; i++) {
+    auto arg_type = env.get_types_before_op(m_my_idx).get(m_expr.get_arg(i).var().reg());
+    TypeSpec desired_type(i == 2 ? "float" : "vector");
+    if (arg_type.typespec() != desired_type) {
+      popped_args.at(i) = cast_form(popped_args.at(i), desired_type, pool, env);
+    }
+  }
+
+  auto new_form = pool.alloc_element<GenericElement>(
+      GenericOperator::make_fixed(FixedOperatorKind::VECTOR_FLOAT_PRODUCT),
+      std::vector<Form*>{popped_args.at(0), popped_args.at(1), popped_args.at(2)});
   result->push_back(new_form);
 }
 
@@ -1430,6 +1511,28 @@ void SimpleExpressionElement::update_from_stack_float_to_int(const Env& env,
   }
 }
 
+namespace {
+GenericElement* allocate_fixed_op(FormPool& pool, FixedOperatorKind kind, Form* op1) {
+  return pool.alloc_element<GenericElement>(GenericOperator::make_fixed(kind), op1);
+}
+}  // namespace
+
+void SimpleExpressionElement::update_from_stack_subu_l32_s7(const Env& env,
+                                                            FormPool& pool,
+                                                            FormStack& stack,
+                                                            std::vector<FormElement*>* result,
+                                                            bool allow_side_effects) {
+  auto var = m_expr.get_arg(0).var();
+  auto arg = pop_to_forms({var}, env, pool, stack, allow_side_effects).at(0);
+  auto type = env.get_types_before_op(var.idx()).get(var.reg()).typespec();
+  if (type != TypeSpec("handle")) {
+    env.func->warnings.general_warning(
+        ".subu (32-bit) used on a {} at idx {}. This probably should be a handle.", type.print(),
+        var.idx());
+  }
+  result->push_back(allocate_fixed_op(pool, FixedOperatorKind::L32_NOT_FALSE_CBOOL, arg));
+}
+
 void SimpleExpressionElement::update_from_stack(const Env& env,
                                                 FormPool& pool,
                                                 FormStack& stack,
@@ -1559,6 +1662,18 @@ void SimpleExpressionElement::update_from_stack(const Env& env,
       break;
     case SimpleExpression::Kind::PCPYLD:
       update_from_stack_pcypld(env, pool, stack, result, allow_side_effects);
+      break;
+    case SimpleExpression::Kind::VECTOR_PLUS:
+      update_from_stack_vector_plus_minus(true, env, pool, stack, result, allow_side_effects);
+      break;
+    case SimpleExpression::Kind::VECTOR_MINUS:
+      update_from_stack_vector_plus_minus(false, env, pool, stack, result, allow_side_effects);
+      break;
+    case SimpleExpression::Kind::VECTOR_FLOAT_PRODUCT:
+      update_from_stack_vector_float_product(env, pool, stack, result, allow_side_effects);
+      break;
+    case SimpleExpression::Kind::SUBU_L32_S7:
+      update_from_stack_subu_l32_s7(env, pool, stack, result, allow_side_effects);
       break;
     default:
       throw std::runtime_error(
@@ -2580,6 +2695,108 @@ void CondWithElseElement::push_to_stack(const Env& env, FormPool& pool, FormStac
 // ShortCircuitElement
 ///////////////////
 
+FormElement* sc_to_handle_get_proc(ShortCircuitElement* elt,
+                                   const Env& env,
+                                   FormPool& pool,
+                                   FormStack& stack) {
+  if (elt->kind != ShortCircuitElement::AND) {
+    return nullptr;
+  }
+
+  if (elt->entries.size() != 2) {
+    return nullptr;
+  }
+
+  // fmt::print("candidate: {}\n", elt->to_string(env));
+
+  constexpr int reg_input_1 = 0;
+  constexpr int reg_input_2 = 1;
+  constexpr int reg_input_3 = 2;
+  constexpr int reg_temp_1 = 10;
+  constexpr int reg_temp_2 = 11;
+  constexpr int reg_temp_3 = 12;
+
+  // check first.
+  auto first_matcher =
+      Matcher::op(GenericOpMatcher::condition(IR2_Condition::Kind::NONZERO),
+                  {Matcher::op(GenericOpMatcher::fixed(FixedOperatorKind::L32_NOT_FALSE_CBOOL),
+                               {Matcher::any_reg(reg_input_1)})});
+
+  auto first_result = match(first_matcher, elt->entries.at(0).condition);
+  if (!first_result.matched) {
+    return nullptr;
+  }
+
+  // auto first_use_of_in = *first_result.maps.regs.at(reg_input_1);
+  // fmt::print("reg1: {}\n", first_use_of_in.to_string(env));
+
+  auto setup_matcher = Matcher::set_var(
+      Matcher::deref(Matcher::any_reg(reg_input_2), false,
+                     {DerefTokenMatcher::string("process"), DerefTokenMatcher::integer(0)}),
+      reg_temp_1);
+
+  auto if_matcher = Matcher::if_no_else(
+      Matcher::op(
+          GenericOpMatcher::fixed(FixedOperatorKind::EQ),
+          {Matcher::deref(Matcher::any_reg(reg_input_3), false, {DerefTokenMatcher::string("pid")}),
+           Matcher::deref(Matcher::any_reg(reg_temp_2), false,
+                          {DerefTokenMatcher::string("pid")})}),
+      Matcher::any_reg(reg_temp_3));
+
+  auto second_matcher = Matcher::begin({setup_matcher, if_matcher});
+
+  auto second_result = match(second_matcher, elt->entries.at(1).condition);
+  if (!second_result.matched) {
+    return nullptr;
+  }
+
+  auto in1 = *first_result.maps.regs.at(reg_input_1);
+  auto in2 = *second_result.maps.regs.at(reg_input_2);
+  auto in3 = *second_result.maps.regs.at(reg_input_3);
+
+  auto in_name = in1.to_string(env);
+  if (in_name != in2.to_string(env)) {
+    return nullptr;
+  }
+
+  if (in_name != in3.to_string(env)) {
+    return nullptr;
+  }
+
+  auto temp_name = second_result.maps.regs.at(reg_temp_1)->to_string(env);
+  if (temp_name != second_result.maps.regs.at(reg_temp_2)->to_string(env)) {
+    return nullptr;
+  }
+
+  if (temp_name != second_result.maps.regs.at(reg_temp_3)->to_string(env)) {
+    return nullptr;
+  }
+
+  const auto& temp_use_def = env.get_use_def_info(*second_result.maps.regs.at(reg_temp_1));
+  if (temp_use_def.use_count() != 2 || temp_use_def.def_count() != 1) {
+    return nullptr;
+  }
+
+  // modify use def:
+  auto* menv = const_cast<Env*>(&env);
+  menv->disable_use(in2);
+  menv->disable_use(in3);
+
+  auto repopped = stack.pop_reg(in1, {}, env, true);
+  // fmt::print("repopped: {}\n", repopped->to_string(env));
+
+  if (!repopped) {
+    repopped = var_to_form(in1, pool);
+  }
+
+  return pool.alloc_element<GenericElement>(
+      GenericOperator::make_function(
+          pool.alloc_single_element_form<ConstantTokenElement>(nullptr, "handle->process")),
+      repopped);
+
+  return nullptr;
+}
+
 void ShortCircuitElement::push_to_stack(const Env& env, FormPool& pool, FormStack& stack) {
   mark_popped();
   if (!used_as_value.value_or(false)) {
@@ -2625,8 +2842,14 @@ void ShortCircuitElement::push_to_stack(const Env& env, FormPool& pool, FormStac
       }
     }
 
+    FormElement* to_push = this;
+    auto as_handle_get = sc_to_handle_get_proc(this, env, pool, stack);
+    if (as_handle_get) {
+      to_push = as_handle_get;
+    }
+
     assert(used_as_value.has_value());
-    stack.push_value_to_reg(final_result, pool.alloc_single_form(nullptr, this), true,
+    stack.push_value_to_reg(final_result, pool.alloc_single_form(nullptr, to_push), true,
                             env.get_variable_type(final_result, false));
     already_rewritten = true;
   }
@@ -2955,6 +3178,15 @@ FormElement* ConditionElement::make_generic(const Env& env,
                                                 casted);
     }
 
+    case IR2_Condition::Kind::LESS_THAN_ZERO_UNSIGNED: {
+      auto casted = make_casts_if_needed(source_forms, types, TypeSpec("uint"), pool, env);
+      auto zero = pool.alloc_single_element_form<SimpleAtomElement>(
+          nullptr, SimpleAtom::make_int_constant(0));
+      casted.push_back(zero);
+      return pool.alloc_element<GenericElement>(GenericOperator::make_fixed(FixedOperatorKind::LT),
+                                                casted);
+    }
+
     case IR2_Condition::Kind::GEQ_ZERO_SIGNED: {
       return make_geq_zero_signed_check_generic(env, pool, source_forms, types);
     }
@@ -3122,10 +3354,14 @@ void ReturnElement::push_to_stack(const Env& env, FormPool& pool, FormStack& sta
   std::vector<FormElement*> new_entries;
   new_entries = rewrite_to_get_var(temp_stack, pool, env.end_var(), env);
 
+  assert(!new_entries.empty());
   return_code->clear();
-  for (auto e : new_entries) {
-    return_code->push_back(e);
+
+  for (int i = 0; i < ((int)new_entries.size()) - 1; i++) {
+    stack.push_form_element(new_entries.at(i), true);
   }
+
+  return_code->push_back(new_entries.back());
   stack.push_form_element(this, true);
 }
 
@@ -3732,6 +3968,130 @@ void StackSpillStoreElement::push_to_stack(const Env& env, FormPool& pool, FormS
   stack.push_form_element(pool.alloc_element<SetFormFormElement>(dst, src), true);
 }
 
+namespace {
+
+/*!
+ * Is the given form an assembly form to load data from a vector to the given vf register?
+ */
+Form* is_load_store_vector_to_reg(const Register& reg,
+                                  FormElement* form,
+                                  bool is_load,
+                                  int* idx_out) {
+  auto as_vf_op = dynamic_cast<VectorFloatLoadStoreElement*>(form);
+  if (!as_vf_op) {
+    return nullptr;
+  }
+
+  if (as_vf_op->is_load() != is_load) {
+    return nullptr;
+  }
+
+  if (as_vf_op->vf_reg() != reg) {
+    return nullptr;
+  }
+
+  // check that we actually got a real vector type, not some other thing that happens to have a
+  // quad.
+  auto& addr_type = as_vf_op->addr_type();
+  if (!addr_type || addr_type != TypeSpec("vector")) {
+    return nullptr;
+  }
+
+  // make sure we load from the right spot, and extract the base.
+  auto loc = as_vf_op->location();
+  auto matcher = Matcher::deref(Matcher::any(0), true, {DerefTokenMatcher::string("quad")});
+  auto mr = match(matcher, loc);
+  if (!mr.matched) {
+    return nullptr;
+  }
+
+  // output the index of the actual store op, for reguse purposes.
+  if (idx_out) {
+    *idx_out = as_vf_op->my_idx();
+  }
+
+  // got it!
+  return mr.maps.forms.at(0);
+}
+
+/*!
+ * Try to convert a form to a regaccess.
+ */
+std::optional<RegisterAccess> form_as_ra(Form* form) {
+  auto as_atom = form_as_atom(form);
+  if (as_atom && as_atom->is_var()) {
+    return as_atom->var();
+  }
+  return {};
+}
+
+bool try_vector_reset_inline(const Env& env,
+                             FormPool& pool,
+                             FormStack& stack,
+                             FormElement* store_element) {
+  // the store
+  int store_idx = -1;
+  auto store = is_load_store_vector_to_reg(Register(Reg::VF, 0), store_element, false, &store_idx);
+  if (!store) {
+    return false;
+  }
+
+  // remove these from the stack.
+  // stack.pop(1);
+
+  // the store here _should_ have failed propagation and just given us a variable.
+  // if this is causing issues, we can run this check before propagating, as well call this from
+  // the function that attempts the pop.
+  auto store_var = form_as_ra(store);
+  if (!store_var) {
+    env.func->warnings.general_warning("Almost found vector reset, but couldn't get store var.");
+    // stack.push_form_element(new_thing->elts().at(0), true);
+    return false;
+  }
+
+  // ignore the store as a use. This will allow the entire vector-! expression to be expression
+  // propagated, if it is appropriate.
+  if (store_var) {
+    auto menv = const_cast<Env*>(&env);
+    menv->disable_use(*store_var);
+  }
+
+  // now try to see if we can pop the first arg (destination vector).
+  bool got_orig = false;
+  RegisterAccess orig;
+  store = repop_passthrough_arg(store, stack, env, &orig, &got_orig);
+
+  // create the actual  form
+  Form* new_thing = pool.alloc_single_element_form<GenericElement>(
+      nullptr,
+      GenericOperator::make_function(
+          pool.alloc_single_element_form<ConstantTokenElement>(nullptr, "vector-reset!")),
+      std::vector<Form*>{store});
+
+  if (got_orig) {
+    // we got a value for the destination.  because we used the special repop passthrough,
+    // we're responsible for inserting a set to set the var that we "stole" from.
+    // We do this through push_value_to_reg, so it can be propagated if needed, but only if
+    // somebody will actually read the output.
+    // to tell, we look at the live out of the store op and the end - the earlier one would of
+    // course be live out always because the store will read it again.
+
+    auto& op_info = env.reg_use().op.at(store_idx);
+    if (op_info.live.find(orig.reg()) == op_info.live.end()) {
+      // nobody reads it, don't bother.
+      stack.push_form_element(new_thing->elts().at(0), true);
+    } else {
+      stack.push_value_to_reg(orig, new_thing, true, TypeSpec("vector"));
+    }
+
+  } else {
+    stack.push_form_element(new_thing->elts().at(0), true);
+  }
+
+  return true;
+}
+}  // namespace
+
 void VectorFloatLoadStoreElement::push_to_stack(const Env& env, FormPool& pool, FormStack& stack) {
   mark_popped();
 
@@ -3740,10 +4100,25 @@ void VectorFloatLoadStoreElement::push_to_stack(const Env& env, FormPool& pool, 
     auto root = loc_as_deref->base();
     auto atom = form_as_atom(root);
     if (atom && atom->get_kind() == SimpleAtom::Kind::VARIABLE) {
-      loc_as_deref->set_base(pop_to_forms({atom->var()}, env, pool, stack, true).at(0));
+      m_addr_type = env.get_variable_type(atom->var(), true);
     }
   }
 
+  auto name = env.func->guessed_name.to_string();
+  // don't find vector-! inside of vector-!.
+  if (!m_is_load && name != "vector-!" && name != "vector+!" && name != "vector-reset!") {
+    if (try_vector_reset_inline(env, pool, stack, this)) {
+      return;
+    }
+  }
+
+  if (loc_as_deref) {
+    auto root = loc_as_deref->base();
+    auto atom = form_as_atom(root);
+    if (atom && atom->get_kind() == SimpleAtom::Kind::VARIABLE) {
+      loc_as_deref->set_base(pop_to_forms({atom->var()}, env, pool, stack, true).at(0));
+    }
+  }
   stack.push_form_element(this, true);
 }